@nalanj 2025-01-03
And this one finished in 5m52s in CI, our new record during this process of optimizing test perf.